Add an environment flag for LILAC_READONLY

+++ src/router_concept.py
@@ -81,6 +81,9 @@ def edit_concept(namespace: str, concept_name: str, change: ConceptUpdate) -> Co
 @router.delete('/{namespace}/{concept_name}')
 def delete_concept(namespace: str, concept_name: str) -> None:
   """Deletes the concept from the database."""
+  if CONFIG.get('LILAC_READONLY', False):
+    raise ValueError('Server is in readonly mode. This disables concept.delete_concept.')
+
   DISK_CONCEPT_DB.remove(namespace, concept_name)
   # Delete concept models from all datasets that are using this concept.
   DISK_CONCEPT_MODEL_DB.remove_all(namespace, concept_name)
