+++ src/router_dataset.py
@@ -6,7 +6,7 @@ from fastapi import APIRouter, Response
 from fastapi.responses import ORJSONResponse
 from pydantic import BaseModel, validator
 
-from .config import data_path
+from .config import CONFIG, data_path
 from .data.dataset import BinaryOp
 from .data.dataset import Column as DBColumn
 from .data.dataset import DatasetManifest, DatasetSettings, FeatureListValue, FeatureValue
@@ -82,6 +82,9 @@ class ComputeSignalOptions(BaseModel):
 @router.delete('/{namespace}/{dataset_name}')
 def delete_dataset(namespace: str, dataset_name: str) -> None:
   """Delete the dataset."""
+  if CONFIG.get('LILAC_READONLY', False):
+    raise ValueError('Server is in readonly mode. This disables dataset.delete_dataset.')
+
   dataset = get_dataset(namespace, dataset_name)
   dataset.delete()
   remove_dataset_from_cache(namespace, dataset_name)
@@ -96,6 +99,8 @@ class ComputeSignalResponse(BaseModel):
 def compute_signal(namespace: str, dataset_name: str,
                    options: ComputeSignalOptions) -> ComputeSignalResponse:
   """Compute a signal for a dataset."""
+  if CONFIG.get('LILAC_READONLY', False):
+    raise ValueError('Server is in readonly mode. This disables dataset.compute_signal.')
 
   def _task_compute_signal(namespace: str, dataset_name: str, options_dict: dict,
                            task_id: TaskId) -> None:
@@ -130,6 +135,9 @@ class DeleteSignalResponse(BaseModel):
 def delete_signal(namespace: str, dataset_name: str,
                   options: DeleteSignalOptions) -> DeleteSignalResponse:
   """Delete a signal from a dataset."""
+  if CONFIG.get('LILAC_READONLY', False):
+    raise ValueError('Server is in readonly mode. This disables dataset.delete_signal.')
+
   dataset = get_dataset(namespace, dataset_name)
   dataset.delete_signal(options.signal_path)
   return DeleteSignalResponse(completed=True)
@@ -293,6 +301,9 @@ def get_settings(namespace: str, dataset_name: str) -> DatasetSettings:
 @router.post('/{namespace}/{dataset_name}/settings', response_model_exclude_none=True)
 def update_settings(namespace: str, dataset_name: str, settings: DatasetSettings) -> None:
   """Get the media for the dataset."""
+  if CONFIG.get('LILAC_READONLY', False):
+    raise ValueError('Server is in readonly mode. This disables dataset.update_settings.')
+
   dataset = get_dataset(namespace, dataset_name)
   dataset.update_settings(settings)
   return None
